Client Side done, now testing for postman

+++ kuhako/src/main/java/com/capstone/kuhako/services/ClientModuleServices/SchedulePaymentRemindersServiceImpl.java
@@ -1,7 +1,10 @@
 package com.capstone.kuhako.services.ClientModuleServices;
 
+import com.capstone.kuhako.models.Client;
+import com.capstone.kuhako.models.ClientModules.DuePayments;
 import com.capstone.kuhako.models.ClientModules.SchedulePaymentReminders;
 import com.capstone.kuhako.repositories.ClientModuleRepository.SchedulePaymentRemindersRepository;
+import com.capstone.kuhako.repositories.ClientRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -12,35 +15,48 @@ public class SchedulePaymentRemindersServiceImpl implements SchedulePaymentRemin
     @Autowired
     private SchedulePaymentRemindersRepository schedulePaymentRemindersRepository;
 
-    public void createSchedulePaymentReminders(SchedulePaymentReminders schedulePaymentReminders){
-//      User author = userRepository.findByUsername(jwtToken.getUsernameFromToken(stringToken));
-        SchedulePaymentReminders newSchedulePaymentReminders = new SchedulePaymentReminders();
-        newSchedulePaymentReminders.setClient(schedulePaymentReminders.getClient());
-        newSchedulePaymentReminders.setPaymentDues(schedulePaymentReminders.getPaymentDues());
-        newSchedulePaymentReminders.setCollectionDate(schedulePaymentReminders.getCollectionDate());
-        newSchedulePaymentReminders.setProductTypeID(schedulePaymentReminders.getProductTypeID());
-        newSchedulePaymentReminders.setReminderTitle(schedulePaymentReminders.getReminderTitle());
-        newSchedulePaymentReminders.setReminderDate(schedulePaymentReminders.getReminderDate());
-        schedulePaymentRemindersRepository.save(newSchedulePaymentReminders);
+    @Autowired
+    private ClientRepository clientRepository;
+    
+    public void createSchedulePaymentReminders(Long clientId,SchedulePaymentReminders schedulePaymentReminders){
+        Client client = clientRepository.findById(clientId).orElse(null);
+        if (client != null) {
+            schedulePaymentReminders.setClient(client);
+            schedulePaymentRemindersRepository.save(schedulePaymentReminders);
+        }
     }
     public Iterable<SchedulePaymentReminders> getSchedulePaymentReminders(){
         return schedulePaymentRemindersRepository.findAll();
     }
 
-    public ResponseEntity deleteSchedulePaymentReminders(Long id){
-        schedulePaymentRemindersRepository.deleteById(id);
-        return new ResponseEntity<>("SchedulePaymentReminders Deleted successfully", HttpStatus.OK);
+    public Iterable<SchedulePaymentReminders> getSchedulePaymentRemindersByClientId(Long clientId) {
+        return schedulePaymentRemindersRepository.findSchedulePaymentRemindersByClientId(clientId);
+    }
+    
+    public ResponseEntity deleteSchedulePaymentReminders(Long clientId,Long id){
+        SchedulePaymentReminders schedulePaymentRemindersToDelete = schedulePaymentRemindersRepository.findById(id).orElse(null);
+
+        if (schedulePaymentRemindersToDelete != null && schedulePaymentRemindersToDelete.getClient().getClient_id().equals(clientId)) {
+            schedulePaymentRemindersRepository.deleteById(id);
+            return new ResponseEntity<>("Schedule Payment Reminders Deleted successfully", HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>("Schedule Payment Reminders not found or does not belong to the specified client", HttpStatus.NOT_FOUND);
+        }
     }
 
-    public ResponseEntity updateSchedulePaymentReminders(Long id, SchedulePaymentReminders schedulePaymentReminders){
-        SchedulePaymentReminders schedulePaymentRemindersForUpdate = schedulePaymentRemindersRepository.findById(id).get();
-        schedulePaymentRemindersForUpdate.setClient(schedulePaymentReminders.getClient());
-        schedulePaymentRemindersForUpdate.setPaymentDues(schedulePaymentReminders.getPaymentDues());
-        schedulePaymentRemindersForUpdate.setCollectionDate(schedulePaymentReminders.getCollectionDate());
-        schedulePaymentRemindersForUpdate.setProductTypeID(schedulePaymentReminders.getProductTypeID());
-        schedulePaymentRemindersForUpdate.setReminderTitle(schedulePaymentReminders.getReminderTitle());
-        schedulePaymentRemindersForUpdate.setReminderDate(schedulePaymentReminders.getReminderDate());
-        schedulePaymentRemindersRepository.save(schedulePaymentRemindersForUpdate);
-        return new ResponseEntity("SchedulePaymentReminders updated successfully", HttpStatus.OK);
+    public ResponseEntity updateSchedulePaymentReminders(Long clientId,Long id, SchedulePaymentReminders schedulePaymentReminders){
+        SchedulePaymentReminders schedulePaymentRemindersForUpdate = schedulePaymentRemindersRepository.findById(id).orElse(null);
+        if (schedulePaymentRemindersForUpdate != null && schedulePaymentRemindersForUpdate.getClient().getClient_id().equals(clientId)) {
+            schedulePaymentRemindersForUpdate.setCollectionDate(schedulePaymentReminders.getCollectionDate());
+            schedulePaymentRemindersForUpdate.setProductTypeID(schedulePaymentReminders.getProductTypeID());
+            schedulePaymentRemindersForUpdate.setReminderTitle(schedulePaymentReminders.getReminderTitle());
+            schedulePaymentRemindersForUpdate.setReminderDate(schedulePaymentReminders.getReminderDate());
+            schedulePaymentRemindersRepository.save(schedulePaymentRemindersForUpdate);
+            return new ResponseEntity<>("SchedulePaymentReminders  Updated successfully", HttpStatus.OK);
+        }
+        return new ResponseEntity<>("Schedule Payment Reminders  not found",HttpStatus.NOT_FOUND);
     }
 }
+
+
+
