+++ kuhako/src/main/java/com/capstone/kuhako/services/ClientModuleServices/TransactionDetailsServiceImpl.java
@@ -1,7 +1,10 @@
 package com.capstone.kuhako.services.ClientModuleServices;
 
+import com.capstone.kuhako.models.Client;
+import com.capstone.kuhako.models.ClientModules.DuePayments;
 import com.capstone.kuhako.models.ClientModules.TransactionDetails;
 import com.capstone.kuhako.repositories.ClientModuleRepository.TransactionDetailsRepository;
+import com.capstone.kuhako.repositories.ClientRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -12,35 +15,48 @@ public class TransactionDetailsServiceImpl implements TransactionDetailsService{
     @Autowired
     private TransactionDetailsRepository transactionDetailsRepository;
 
-    public void createTransactionDetails(TransactionDetails transactionDetails){
-//      User author = userRepository.findByUsername(jwtToken.getUsernameFromToken(stringToken));
-        TransactionDetails newTransactionDetails = new TransactionDetails();
-        newTransactionDetails.setClient(transactionDetails.getClient());
-        newTransactionDetails.setMonthlyCollectible(transactionDetails.getMonthlyCollectible());
-        newTransactionDetails.setFullCashPrice(transactionDetails.getFullCashPrice());
-        newTransactionDetails.setInterestRate(transactionDetails.getInterestRate());
-        newTransactionDetails.setDueDate(transactionDetails.getDueDate());
-        newTransactionDetails.setTotalPayment(transactionDetails.getTotalPayment());
-        transactionDetailsRepository.save(newTransactionDetails);
+    @Autowired
+    private ClientRepository clientRepository;
+    
+    public void createTransactionDetails(Long clientId,TransactionDetails transactionDetails){
+        Client client = clientRepository.findById(clientId).orElse(null);
+        if (client != null) {
+            transactionDetails.setClient(client);
+            transactionDetailsRepository.save(transactionDetails);
+        }
     }
     public Iterable<TransactionDetails> getTransactionDetails(){
         return transactionDetailsRepository.findAll();
     }
 
-    public ResponseEntity deleteTransactionDetails(Long id){
-        transactionDetailsRepository.deleteById(id);
-        return new ResponseEntity<>("TransactionDetails Deleted successfully", HttpStatus.OK);
+    public Iterable<TransactionDetails> getTransactionDetailsByClientId(Long clientId) {
+        return transactionDetailsRepository.findTransactionDetailsByClientId(clientId);
+    }
+    public ResponseEntity deleteTransactionDetails(Long clientId,Long id){
+
+        TransactionDetails transactionDetailsToDelete = transactionDetailsRepository.findById(id).orElse(null);
+
+        if (transactionDetailsToDelete != null && transactionDetailsToDelete.getClient().getClient_id().equals(clientId)) {
+            transactionDetailsRepository.deleteById(id);
+            return new ResponseEntity<>("Transaction Details Deleted successfully", HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>("Transaction Details not found or does not belong to the specified client", HttpStatus.NOT_FOUND);
+        }
     }
 
-    public ResponseEntity updateTransactionDetails(Long id, TransactionDetails transactionDetails){
-        TransactionDetails transactionDetailsForUpdate = transactionDetailsRepository.findById(id).get();
-        transactionDetailsForUpdate.setClient(transactionDetails.getClient());
-        transactionDetailsForUpdate.setMonthlyCollectible(transactionDetails.getMonthlyCollectible());
-        transactionDetailsForUpdate.setFullCashPrice(transactionDetails.getFullCashPrice());
-        transactionDetailsForUpdate.setInterestRate(transactionDetails.getInterestRate());
-        transactionDetailsForUpdate.setDueDate(transactionDetails.getDueDate());
-        transactionDetailsForUpdate.setTotalPayment(transactionDetails.getTotalPayment());
-        transactionDetailsRepository.save(transactionDetailsForUpdate);
-        return new ResponseEntity("TransactionDetails updated successfully", HttpStatus.OK);
+    public ResponseEntity updateTransactionDetails(Long clientId,Long id, TransactionDetails transactionDetails){
+        TransactionDetails transactionDetailsForUpdate = transactionDetailsRepository.findById(id).orElse(null);
+        if (transactionDetailsForUpdate != null && transactionDetailsForUpdate.getClient().getClient_id().equals(clientId)) {
+            transactionDetailsForUpdate.setMonthlyCollectible(transactionDetails.getMonthlyCollectible());
+            transactionDetailsForUpdate.setFullCashPrice(transactionDetails.getFullCashPrice());
+            transactionDetailsForUpdate.setInterestRate(transactionDetails.getInterestRate());
+            transactionDetailsForUpdate.setDueDate(transactionDetails.getDueDate());
+            transactionDetailsForUpdate.setTotalPayment(transactionDetails.getTotalPayment());
+            transactionDetailsRepository.save(transactionDetailsForUpdate);
+            return new ResponseEntity("Transaction Details updated successfully", HttpStatus.OK);
+        }
+        return new ResponseEntity<>("Transaction Details not found",HttpStatus.NOT_FOUND);
     }
+
+
 }
